requirements.txt
requests>=2.31
ultralytics>=8.0
pyyaml>=6.0
numpy>=1.26
opencv-python>=4.8
pillow>=10.0
pytesseract>=0.3.10
easyocr>=1.7
transformers>=4.38
torch>=2.1
streamlit>=1.34
